[bash] Fix not found problem

+++ after_deploy/03_nvim.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
 # link ~/.config/nvim/after -> ~/.vim/after
-cd ~/.config/nvim
-ln -sf ~/.vim/after .
+if [ -e ~/.config/nvim ]; then
+  cd ~/.config/nvim
+  ln -sf ~/.vim/after .
+fi
